Minor: Refactor Services structure
