+++ src/app/services/index.ts
@@ -1,9 +1,9 @@
-export * from './scheduling.service';
+export * from './scheduling/scheduling.service';
 export * from './user/user.service';
 export * from './council-group/council-group.service';
-export * from './checklist.service';
+export * from './checklist/checklist.service';
 export * from './authentication/authentication.service';
 export * from './alert/alert.service';
-export * from './school.service';
+export * from './school/school.service';
 export * from './ibge/ibge.service';
-export * from './services-utilities.service';
+export * from './services-utilities/services-utilities.service';
